# docs/public-api.openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Renewlet Public API",
    "version": "1.0.0",
    "description": "Read-only API for subscriptions, status summaries, and upcoming due items. All endpoints require an `Authorization: Bearer rlt_*` token created in Renewlet settings.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/zhiyingzzhou/renewlet/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin Renewlet deployment."
    }
  ],
...
  "paths": {
    "/api/public/v1/me": {
      "get": {
        "tags": [
          "Public API"
        ],
        "operationId": "getPublicApiMe",
        "summary": "Inspect the current Public API token",
        "description": "Returns the read scope attached to the bearer token. It does not expose the token owner or token hash.",
        "security": [
          {
            "PublicApiBearer": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiMeResponse"
                },
                "examples": {
                  "success": {
                    "value": {
                      "ok": true,
                      "data": {
                        "scopes": [
                          "read"
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, deleted, or unauthorized Public API token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorResponse"
                },
                "examples": {
                  "error": {
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "message": "Unauthorized"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/public/v1/subscriptions": {
      "get": {
        "tags": [
          "Public API"
        ],
        "operationId": "listPublicApiSubscriptions",
        "summary": "List subscriptions",
        "description": "Returns an owner-scoped page of subscriptions ordered by Renewlet's Public API cursor contract.",
        "security": [
          {
            "PublicApiBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size. The maximum is 100.",
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "example": 50
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Opaque cursor returned as `data.nextCursor` from the previous page.",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512
            },
            "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTAxVDAwOjAwOjAwLjAwMFoiLCJpZCI6InN1YiJ9"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiSubscriptionsListResponse"
                },
                "examples": {
                  "success": {
                    "value": {
                      "ok": true,
                      "data": {
                        "subscriptions": [
                          {
                            "id": "sub_01HZY7YQ7EXAMPLE",
                            "name": "GitHub",
                            "price": "12",
                            "currency": "USD",
                            "billingCycle": "monthly",
                            "category": "developer_tools",
                            "status": "active",
                            "pinned": false,
                            "publicHidden": false,
                            "startDate": "2026-01-01",
                            "nextBillingDate": "2026-09-01",
                            "autoRenew": false,
                            "autoCalculateNextBillingDate": true,
                            "tags": [
                              "dev"
                            ],
                            "reminderDays": 3,
                            "repeatReminderEnabled": false,
                            "repeatReminderInterval": "1h",
                            "repeatReminderWindow": "72h",
                            "extra": {},
                            "createdAt": "2026-08-01T00:00:00.000Z",
                            "updatedAt": "2026-08-01T00:00:00.000Z"
                          }
                        ],
                        "nextCursor": null,
                        "total": 1
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorResponse"
                },
                "examples": {
                  "error": {
                    "value": {
                      "error": {
                        "code": "INVALID_QUERY",
                        "message": "Invalid request parameters"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, deleted, or unauthorized Public API token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorResponse"
                },
                "examples": {
                  "error": {
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "message": "Unauthorized"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/public/v1/subscriptions/{id}": {
      "get": {
        "tags": [
          "Public API"
        ],
        "operationId": "getPublicApiSubscription",
        "summary": "Read one subscription",
        "description": "Returns one owner-scoped subscription by id. Foreign or missing ids return the same not found shape.",
        "security": [
          {
            "PublicApiBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Subscription id from the list endpoint.",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 120,
              "description": "Renewlet subscription id returned by the subscriptions list endpoint."
            },
            "example": "sub_01HZY7YQ7EXAMPLE"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiSubscriptionResponse"
                },
                "examples": {
                  "success": {
                    "value": {
                      "ok": true,
                      "data": {
                        "subscription": {
                          "id": "sub_01HZY7YQ7EXAMPLE",
                          "name": "GitHub",
                          "price": "12",
                          "currency": "USD",
                          "billingCycle": "monthly",
                          "category": "developer_tools",
                          "status": "active",
                          "pinned": false,
                          "publicHidden": false,
                          "startDate": "2026-01-01",
                          "nextBillingDate": "2026-09-01",
                          "autoRenew": false,
                          "autoCalculateNextBillingDate": true,
                          "tags": [
                            "dev"
                          ],
                          "reminderDays": 3,
                          "repeatReminderEnabled": false,
                          "repeatReminderInterval": "1h",
                          "repeatReminderWindow": "72h",
                          "extra": {},
                          "createdAt": "2026-08-01T00:00:00.000Z",
                          "updatedAt": "2026-08-01T00:00:00.000Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, deleted, or unauthorized Public API token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorResponse"
                },
                "examples": {
                  "error": {
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "message": "Unauthorized"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist for the token owner.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorResponse"
                },
                "examples": {
                  "error": {
                    "value": {
                      "error": {
                        "code": "NOT_FOUND",
                        "message": "Not found"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "PublicApiBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "rlt_*",
        "description": "Renewlet Public API token. Query parameter tokens and browser session cookies are rejected."
      }
    },
    "schemas": {
      "PublicApiErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1
              },
              "message": {
                "type": "string"
              },
              "details": {},
              "requestId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "PublicApiMeResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "object",
            "properties": {
              "scopes": {
                "minItems": 1,
                "maxItems": 1,
                "type": "array",
                "items": {
                  "type": "string",
                  "const": "read"
                }
              }
            },
            "required": [
              "scopes"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "ok",
          "data"
        ],
        "additionalProperties": false
      },
...
    }
  }
}